pfz_system/requirements.txt
# PFZ System Dependencies
# 潛在漁場預測系統

# Core
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# HTTP & API
requests>=2.28.0

# Configuration
python-dotenv>=1.0.0

# Date/Time
python-dateutil>=2.8.0

# Type hints (dev)
typing-extensions>=4.5.0

# Optional: Visualization
# matplotlib>=3.7.0
# plotly>=5.14.0
# cartopy>=0.21.0

# API Server
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Optional: Linting
# black>=23.3.0
# flake8>=6.0.0
# mypy>=1.3.0
